Background and OneVsThree
1. Fixed the board, so its drawn properly (the "bases" are now as they
are supposed to be).

2. Added lots of stuff in the OneVsThree class. Going to use this for
adding the logic, before implementing the superclass. Reason for this is
simply because it's easier to test with one class. Once this class is
functionable, we can move code that's going to be used regardless of
game mode, to the superclass.

What I did:
Created the path array. This array now contains a rectangle for _every
single_ possible location on the map. This is, in total, 92 fields. Will
create illustration to show which array field goes where, to make it
easier for development.

Didn't bother with for-loops etc now. Going to fix that later. Code is
unnecessary long as it is right now. Reason for drawing a piece in every
single position, is to see that it's correct. The drawing statements can
be removed now.

Next up is to create paths for each color. Once that's done, we can
start adding logic to the movement system. Create methods that allows
movement, and a system that knows whose turn it is etc.

I couldn't sleep, so I spent all night doing this, instead of looking at
the ceiling for 5 hours.