Last patch for the night
Changed some logic. AI is now an interface, where we can create multiple
AI's.

Stuff that should be fixed tomorrow, as a last bug-fix:
make it impossible to "waste" moves, by clicking a piece inside your
base when you have a 4.
If you roll a 6, you get an extra move.